refactor(formatter): split expression.rs into expression/ module
Pure move: function bodies and comments are moved verbatim; only `use`
statements and item visibility (`pub(super)` for cross-module helpers)
were adjusted.
